kvm: It's no longer required to have VNC listen in 0.0.0.0
In both the master and 4.3 branch there is a patch which sets
the VNC address to the private IP of the hypervisor and during
migration this will be changed to the private IP of the destination
hypervisor.

Binding VNC on 0.0.0.0 is a potential security issue with the hypervisors
also have a public IP address and the VNC ports are not firewalled.